Don't try to remove directories that don't exist

+++ stash-merritt/lib/stash/merritt/zip_package.rb
@@ -33,6 +33,7 @@ module Stash
       end
 
       def cleanup!
+        return unless File.exists?(workdir)
         FileUtils.remove_dir(workdir)
         @zipfile = nil
       end
